requirements.txt
flask
flask-cors
transformers
torch
pillow
requests
